refactor: 반경 원의 범위 계산 로직과 cctv 데이터 정렬 로직 cctv 도메인으로 이동
