+++ src/main/java/service/tmbmbackend/service/Impl/CCTVServiceImpl.java
@@ -1,13 +1,11 @@
 package service.tmbmbackend.service.Impl;
 
-import java.util.Comparator;
+import static service.tmbmbackend.entity.CCTV.calculateCircleRadius;
+import static service.tmbmbackend.entity.CCTV.sortCCTVListByDistance;
+
 import java.util.List;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
-import org.locationtech.jts.geom.Coordinate;
-import org.locationtech.jts.geom.Geometry;
-import org.locationtech.jts.geom.Point;
-import org.locationtech.jts.util.GeometricShapeFactory;
 import org.springframework.data.domain.Page;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -28,36 +26,13 @@ public class CCTVServiceImpl implements CCTVService {
     @Override
     public Response getAllCCTVs(CCTVZoneRequest cctvZoneRequest) {
         Page<CCTV> cctvs = cctvRepository.findWithin(
-                calculateCircleRadius(cctvZoneRequest.getX(), cctvZoneRequest.getY(), cctvZoneRequest.getRadius()), cctvZoneRequest.getPageRequest());
-        List<DataResponse> dataResponses = CCTVDataResponse.toList(
+                calculateCircleRadius(cctvZoneRequest.getX(), cctvZoneRequest.getY(),
+                        cctvZoneRequest.getRadius()), cctvZoneRequest.getPageRequest());
+
+        CCTVMetaResponse cctvMetaResponse = new CCTVMetaResponse(cctvZoneRequest, cctvs.hasPrevious(), cctvs.hasNext());
+        List<DataResponse> cctvDataResponses = CCTVDataResponse.toList(
                 sortCCTVListByDistance(cctvZoneRequest.getX(), cctvZoneRequest.getY(),
                         cctvs.getContent().stream().collect(Collectors.toList())));
-        return new Response(new CCTVMetaResponse(cctvZoneRequest, cctvs.hasPrevious(), cctvs.hasNext()), dataResponses);
-    }
-
-    private static List<CCTV> sortCCTVListByDistance(double x, double y, List<CCTV> cctvList) {
-        cctvList.sort(Comparator.comparingDouble(c -> calculateDistance(x, y, c.getPoint())));
-        return cctvList;
-    }
-
-    private static double calculateDistance(double x, double y, Point point) {
-        double earthRadius = 6371;
-
-        double lat1 = Math.toRadians(x);
-        double lon1 = Math.toRadians(y);
-        double lat2 = Math.toRadians(point.getX());
-        double lon2 = Math.toRadians(point.getY());
-
-        return earthRadius * Math.acos(Math.sin(lat1) * Math.sin(lat2) + Math.cos(lat1) * Math.cos(lat2) * Math.cos(lon1 - lon2));
-    }
-
-    private Geometry calculateCircleRadius(double x, double y, double radius) {
-        double meterToCoordinate = (radius / 100000);
-
-        GeometricShapeFactory shapeFactory = new GeometricShapeFactory();
-        shapeFactory.setNumPoints(32);
-        shapeFactory.setCentre(new Coordinate(x, y));
-        shapeFactory.setSize(meterToCoordinate * 2);
-        return shapeFactory.createCircle();
+        return new Response(cctvMetaResponse, cctvDataResponses);
     }
 }
